Refactor group tests and move helper functions inline
